Fix collision detection.
When the ball intersects with two bricks after a time step, a function
was called to calculate the overlap with each of these bricks. Due to a
bug, the calculation always returned NaN and the ball always bounced off
of the second brick (independent from the overlap). This bug has now
been fixed. The bug caused the ball to frequently change direction on
a bounce. This issue is now fixed.

BUG: 328812
FIXED-IN: 16.08
REVIEW: 128472