KBreakout: Fix brick count
Description:
As described in bug 328811, some levels clear before all bricks have been destroyed. This happes for example in Level 6, when invisible bricks are destroyed by the "Burning Ball". KBreakout keeps count of the remaining bricks. Indestructable and invisible bricks do not contribute to the remaining bricks. When bricks are destroyed, the remaining brick counter is decreased accordingly. However, when an invisible or indestructable brick explodes, the remaining brick counter is decreased even though these bricks did not contribute to the number of remaining bricks.

In the patch, the remaining brick counter is increased if an invisible or indestructible brick is set to explode.

Testing Done:
Print out number of remaining bricks.

a) Before the patch: Confirm that number of remaining bricks is wrong.
b) After the patch: Confirm that number of remaining bricks is correct. Confirm that levels no longer clear prematurely.

REVIEW: 128556